+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/LastReadingModel.kt
@@ -2,7 +2,7 @@ package com.yeobaek.data.model
 
 data class LastReadingModel(
     val book: BookModel,
-    val clubId: Int,
+    val clubId: Long,
     val clubName: String,
     val lastReadAt: String,
     val lastReadPassageSequence: Int,
