+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/repository/GroupRepository.kt
@@ -5,7 +5,7 @@ import com.yeobaek.data.model.GroupModel
 
 interface GroupRepository {
     suspend fun getGroups(): List<GroupModel>
-    suspend fun createGroup(groupName: String, bookId: Int)
+    suspend fun createGroup(groupName: String, bookId: Long)
     suspend fun joinGroup(joinCode: String)
-    suspend fun getGroupDetail(groupId: Int): GroupDetailModel
+    suspend fun getGroupDetail(groupId: Long): GroupDetailModel
 }
