+++ android/shared/src/commonMain/kotlin/com/yeobaek/feature/reader/PassageCommentSheetUiState.kt
@@ -3,11 +3,11 @@ package com.yeobaek.feature.reader
 import com.yeobaek.feature.reader.model.PassageCommentUiModel
 
 data class PassageCommentSheetUiState(
-    val passageId: Int,
+    val passageId: Long,
     val comments: List<PassageCommentUiModel> = emptyList(),
     val input: String = "",
-    val editingCommentId: Int? = null,
-    val deletingCommentId: Int? = null,
+    val editingCommentId: Long? = null,
+    val deletingCommentId: Long? = null,
     val isLoading: Boolean = false,
     val isSubmitting: Boolean = false,
     val isDeleting: Boolean = false,
